PostGIS / Replace deprecated GeomFromText by ST_GeomFromText.
Fix for https://github.com/talend-spatial/talend-spatial/issues/82

See https://postgis.net/docs/reference.html

```
	
PostGIS has begun a transition from the existing naming convention to an SQL-MM-centric convention. As a result, most of the functions that you know and love have been renamed using the standard spatial type (ST) prefix. Previous functions are still available, though are not listed in this document where updated functions are equivalent. The non ST_ functions not listed in this documentation are deprecated and will be removed in a future release so STOP USING THEM.
```